requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.114.1
greenlet==3.1.0
h11==0.14.0
idna==3.8
importlib_metadata==8.5.0
platformdirs==4.3.2
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
yapf==0.40.2
zipp==3.20.1
gunicorn==23.0.0
